scripts/requirementscnn.txt
pandas>=1.3.0
openpyxl>=3.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
torch>=1.13.0
torchvision>=0.14.0
Pillow>=8.2.0
pyyaml>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.1
tensorboard>=2.9.0
